+++ src/services/strategyArchive/index.ts
@@ -1,5 +1,5 @@
 export * from './types';
-export { STRATEGY_ARCHIVE, STRATEGY_ARCHIVE_PLANNED, getStrategyDefinition } from './registry';
+export { STRATEGY_ARCHIVE, STRATEGY_ARCHIVE_PLANNED, STRATEGY_ARCHIVE_TOTAL_ROWS, getStrategyDefinition } from './registry';
 export { reproduce, splitFor } from './engine/reproductionEngine';
 export { computeRMetrics, tradesDigest } from './engine/rMetrics';
 export { ohlcvToArchiveCandles, archiveCandlesToOhlcv, detectTimestampUnit, toMs, validateSeries } from './shared/candleAdapter';
@@ -8,3 +8,7 @@ export {
   V30_DEFINITION, V30_SOURCE_RESULTS, V30_REPRODUCTION_EVIDENCE, V30_REPRODUCED_RESULTS, V30_CAVEATS_RU, V30_RULES_RU, V30_DISCREPANCIES, V30_COMMITS, V30_SOURCE_PINS,
 } from './definitions/v3_0-htf-liquidation-trap/definition';
 export { V30_CONSTANTS } from './definitions/v3_0-htf-liquidation-trap/v30Core';
+export {
+  V31_DEFINITION, V31_SOURCE_RESULTS, V31_REPRODUCTION_EVIDENCE, V31_REPRODUCED_RESULTS, V31_CAVEATS_RU, V31_RULES_RU, V31_DISCREPANCIES, V31_COMMITS, V31_SOURCE_PINS, V31_VARIANTS,
+} from './definitions/v3_1-htf-trend-pullback/definition';
+export { V31_CONSTANTS } from './definitions/v3_1-htf-trend-pullback/v31Core';